Refac: overhaul of game_scanner
