+++ src/core/game_scanner.py
@@ -1,7 +1,5 @@
-import json
+
 import os
-import re
-import requests
 
 import vdf
 import yaml
@@ -13,120 +11,20 @@ from core.tools import  write_yaml, load_yaml, slugify
 from typing import List, Dict, Optional, Any
 #from platforms.steam import get_library_paths, get_steam_base_dir, scan_libraries
 from platforms import steam
+from platforms import heroic
 
-
-
-
-
-def get_heroic_config_paths() -> Dict[str, Optional[str]]:
-    paths = {"epic": None, "gog": None}
-    
-    # Epic
-    epic_flatpak = os.path.expanduser("~/.var/app/com.heroicgameslauncher.hgl/config/heroic/legendaryConfig/legendary/installed.json")
-    epic_native = os.path.expanduser("~/.config/heroic/legendaryConfig/legendary/installed.json")
-    # Gives to "epic" key either flatpak path or native path
-    paths["epic"] = epic_flatpak if os.path.exists(epic_flatpak) else (epic_native if os.path.exists(epic_native) else None)
-
-    # GOG
-    gog_flatpak = os.path.expanduser("~/.var/app/com.heroicgameslauncher.hgl/config/heroic/gog_store/installed.json")
-    gog_native = os.path.expanduser("~/.config/heroic/gog_store/installed.json")
-    # Gives to "gog" key either flatpak path or native path
-    paths["gog"] = gog_flatpak if os.path.exists(gog_flatpak) else (gog_native if os.path.exists(gog_native) else None)
-
-    # Returns the paths in a dictionary
-    return paths
-
-def find_game_art(app_id: str | int, platform: str) -> dict:
-    art = {"hero": None, "poster": None}
-    if not app_id: return None
-    if platform == "heroic-epic":
-        paths = download_heroic_assets(app_id, platform)
-        art["poster"] = paths.get("art_square")
-        art["hero"] = paths.get("art_hero")
-        return art if art else None
-    elif platform == "heroic-gog":
-        paths = download_heroic_assets(app_id, platform)
-        art["poster"] = paths.get("art_square")
-        art["hero"] = paths.get("art_hero")
-        return art if art else None
-    return art
-
-def scan_heroic_epic_game(yaml_data, yaml_path, game_title, installed_epic):
-    for app_id, game_info in installed_epic.items():
-        if slugify(game_info.get("title", "")) == slugify(game_title):
-            game_path = game_info.get("install_path", "")
-            yaml_data["platform"] = "heroic-epic"
-            yaml_data["game_path"] = game_path
-            write_yaml(yaml_data, yaml_path)
-            
-            return {
-                "name": game_title,
-                "img": find_game_art(app_id, "heroic-epic"),
-                "path": game_path,
-                "app_id": app_id,
-                "platform": "heroic-epic",
-                "game_config_path": yaml_path
-            }
-    return None
-
-def scan_heroic_gog_game(yaml_data, yaml_path, game_title, installed_gog):
-    if not yaml_data.get("gog_id"):
-        return None
-        
-    for game_info in installed_gog.get("installed", []):
-        if slugify(game_info.get("appName", "")) == slugify(str(yaml_data["gog_id"])):
-            game_path = game_info.get("install_path", "")
-            yaml_data["platform"] = "heroic-gog"
-            yaml_data["game_path"] = game_path
-            write_yaml(yaml_data, yaml_path)
-            
-            return {
-                "name": game_title,
-                "img": find_game_art(yaml_data["gog_id"], "heroic-gog"),
-                "path": game_path,
-                "app_id": yaml_data["gog_id"],
-                "platform": "heroic-gog",
-                "game_config_path": yaml_path
-            }
-    return None
 
 def scan_all_games(game_configs_dir):
     matches = []
     steam_base = steam.get_steam_base_dir()
 
-    game_libraries = []
-
     user_config_dir = os.path.join(GLib.get_user_data_dir(), 'nomm', 'user_config.yaml')
     user_config = load_yaml(user_config_dir)
-    found_libs = set(user_config.get("library_paths", []))
-    
-    # Update Steam Libraries
-    if steam_base:
-        found_libs = set(steam.get_library_paths(os.path.join(steam_base, "config/libraryfolders.vdf")))
-        if found_libs:
-            update_user_config("library_paths", sorted(list(found_libs)))
 
-    game_libraries.extend(list(found_libs))
-    print(f"Game libraries detected: {str(game_libraries)}")
-
-    # Pre-load Heroic Libraries
-    heroic_paths = get_heroic_config_paths()
-    installed_epic = {}
-    installed_gog = {}
-    
-    if heroic_paths["epic"]:
-        try:
-            with open(heroic_paths["epic"], 'r', encoding='utf-8') as f:
-                installed_epic = json.load(f)
-        except Exception as e:
-            print(f"Error loading Epic JSON: {e}")
-            
-    if heroic_paths["gog"]:
-        try:
-            with open(heroic_paths["gog"], 'r', encoding='utf-8') as f:
-                installed_gog = json.load(f)
-        except Exception as e:
-            print(f"Error loading GOG JSON: {e}")
+    # Pre-load Libraries
+    steam_libraries = steam.get_library_paths(steam_base) # list with paths to Steam libraries
+    epic_library = heroic.get_epic_library() # dict with paths to individual games
+    gog_library = heroic.get_gog_library() # dict with paths to individual games
 
     if not os.path.exists(game_configs_dir):
         print(f"Configs directory not found at {game_configs_dir}")
@@ -134,143 +32,52 @@ def scan_all_games(game_configs_dir):
 
     heroic_game_paths = []
 
-    # Scan each config
+    # Scan each game config
     for filename in os.listdir(game_configs_dir):
         if not filename.lower().endswith((".yaml", ".yml")):
             continue
-            
+
         yaml_path = os.path.join(game_configs_dir, filename)
         try:
             with open(yaml_path, 'r', encoding='utf-8') as f:
                 yaml_data = yaml.safe_load(f) or {}
         except Exception as e:
-            print(f"Error processing {filename}: {e}")
-
-        if not yaml_data.get("name") or yaml_data.get("mods_path") is None:
-            print("Yaml game file malformed, skipping...")
+            print(f"[!] Error processing {filename}: {e}, skipping")
             continue
-        
+
+        if not yaml_data.get("name") or "mods_path" not in yaml_data:
+            print("[!] Missing required information in YAML file, skipping...")
+            continue
+
         game_title = yaml_data["name"]
 
         # Scan Steam
-        if found_libs:
-            match = steam.scan_libraries(yaml_data, yaml_path, game_title, found_libs, steam_base)
+        if steam_libraries:
+            match = steam.find_game(yaml_data, yaml_path, game_title, steam_libraries, steam_base)
             if match:
                 matches.append(match)
                 continue
 
         # Scan Heroic Epic
-        if installed_epic:
-            match = scan_heroic_epic_game(yaml_data, yaml_path, game_title, installed_epic)
+        if epic_library:
+            match = heroic.find_epic_game(yaml_data, yaml_path, game_title, epic_library)
             if match:
                 matches.append(match)
                 heroic_game_paths.append(match["path"])
                 continue
 
-        # Scan Heroic Gog
-        if installed_gog:
-            match = scan_heroic_gog_game(yaml_data, yaml_path, game_title, installed_gog)
+        # Scan Heroic GOG
+        if gog_library:
+            match = heroic.find_gog_game(yaml_data, yaml_path, game_title, gog_library)
             if match:
                 matches.append(match)
                 heroic_game_paths.append(match["path"])
                 continue
     
-    heroic_library_paths = obtain_heroic_libraries(heroic_game_paths)
+    heroic_libraries = heroic.obtain_heroic_libraries(heroic_game_paths)
 
-    game_libraries = game_libraries + heroic_library_paths
+    game_libraries = steam_libraries + heroic_libraries
+    print(f"Game libraries detected: {str(game_libraries)}")
+    update_user_config("library_paths", sorted(game_libraries))
 
     return matches, game_libraries
-
-def obtain_heroic_libraries(game_paths: list) -> list:
-    """Takes a list of unique game paths and attempts to extrapolate a list of library directories"""
-    directory_paths = []
-    for path in game_paths:
-        if os.path.dirname(path) not in directory_paths:
-            directory_paths.append(os.path.dirname(path))
-    return directory_paths
-    
-# Grabs the assets from heroic games launcher such as banner and game image
-# TODO: Needs to be cleaned
-def download_heroic_assets(appName: str, platform: str):
-    if isinstance(appName, list):
-        appName = str(appName[0])
-    else:
-        appName = str(appName)
-
-    json_path = os.path.expanduser("~/.var/app/com.heroicgameslauncher.hgl/config/heroic/store/download-manager.json") # flatpak
-    if not os.path.exists(json_path):
-        json_path = os.path.expanduser("~/.config/heroic/store/download-manager.json") # not flatpak
-
-    if isinstance(appName, list):
-        appName = appName[0]
-    
-    cache_base = os.path.join(GLib.get_user_data_dir(), "nomm", "image-cache", f"{platform}", f"{appName}")
-    
-    if os.path.exists(cache_base):
-        existing_files = {}
-        for entry in os.listdir(cache_base):
-            if entry.startswith("art_square"):
-                existing_files["art_square"] = os.path.join(cache_base, entry)
-            elif entry.startswith("art_hero"):
-                existing_files["art_hero"] = os.path.join(cache_base, entry)
-        
-        if "art_square" in existing_files:
-            print(f"Using cached assets for {appName}")
-            return existing_files
-
-    if not os.path.exists(json_path):
-        print(f"Heroic config not found at {json_path}")
-        return None
-
-    try:
-        with open(json_path, 'r') as f:
-            data = json.load(f)
-        
-        finished_apps = data.get("finished", [])
-        target_info = None
-
-        for entry in finished_apps:
-            params = entry.get("params", {})
-            game_info = params.get("gameInfo", {})
-            
-            # Match by internal appName (e.g., 'Curry') or title (e.g., 'ABZÛ')
-            if params.get("appName") == appName or game_info.get("title") == appName:
-                target_info = game_info
-                break
-        
-        if not target_info:
-            return None
-
-        urls = {
-            "art_square": target_info.get("art_square"),
-            "art_hero": target_info.get("art_background") or target_info.get("art_cover")
-        }
-
-        os.makedirs(cache_base, exist_ok=True)
-        downloaded_paths = {}
-
-        for key, url in urls.items():
-            if not url:
-                continue
-                
-            ext = os.path.splitext(url)[1] if "." in url.split("/")[-1] else ".jpg"
-            # Ensure extensions like .jpg?foo=bar are cleaned
-            if "?" in ext: ext = ext.split("?")[0]
-            
-            local_path = os.path.join(cache_base, f"{key}{ext}")
-
-            try:
-                r = requests.get(url, timeout=15)
-                if r.status_code == 200:
-                    with open(local_path, 'wb') as f:
-                        f.write(r.content)
-                    downloaded_paths[key] = local_path
-                    print(f"Downloaded: {local_path}")
-            except Exception as e:
-                print(f"Error downloading {key}: {e}")
-
-        return downloaded_paths
-
-    except Exception as e:
-        print(f"Failed to process Heroic JSON: {e}")
-        return None
